Move test from runtime-common shared tests to realm-server directly
The test only runs in realm-server, so there's no need for the
SharedTests indirection through runtime-common.
